introduction do property base test

+++ go_with_test/roman_numeral/roman_numeral.go
@@ -3,7 +3,7 @@ package romannumeral
 import "strings"
 
 type RomanNumeral struct {
-	Value  int
+	Value  uint16
 	Symbol string
 }
 
@@ -23,7 +23,7 @@ var allRomanNumerals = []RomanNumeral{
 	{1, "I"},
 }
 
-func ConvertToRoman(arabic int) string {
+func ConvertToRoman(arabic uint16) string {
 
 	var result strings.Builder
 
@@ -55,7 +55,7 @@ func ConvertToRoman(arabic int) string {
 	return result.String()
 }
 
-func ConvertToArabic(roman string) int {
+func ConvertToArabic(roman string) uint16 {
 	// if roman == "III" {
 	// 	return 3
 	// }
@@ -67,7 +67,7 @@ func ConvertToArabic(roman string) int {
 	// 	total++
 	// }
 
-	var arabic = 0
+	var arabic = uint16(0)
 	for _, numeral := range allRomanNumerals {
 		for strings.HasPrefix(roman, numeral.Symbol) {
 			arabic += numeral.Value
